Rename files for `Style/Strip` cop
Follow up of https://github.com/rubocop-hq/rubocop/commit/ca0e4d3ed3b1176ce51b587a49bfe7a21d72e288.

This PR fixes the following RuboCop offense.

```console
bundle exec rake internal_investigation

(snip)

Offenses:

spec/rubocop/cop/style/lstrip_rstrip_spec.rb:3:1: C: RSpec/FilePath:
Spec path should end with rubocop/cop/style/strip*_spec.rb.
RSpec.describe RuboCop::Cop::Style::Strip do
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

1198 files inspected, 1 offense detected
RuboCop failed!
Exited with code 1
```

https://circleci.com/gh/rubocop-hq/rubocop/43960
