+++ lib/rubocop.rb
@@ -458,7 +458,6 @@ require_relative 'rubocop/cop/style/inline_comment'
 require_relative 'rubocop/cop/style/ip_addresses'
 require_relative 'rubocop/cop/style/lambda'
 require_relative 'rubocop/cop/style/lambda_call'
-require_relative 'rubocop/cop/style/lstrip_rstrip'
 require_relative 'rubocop/cop/style/line_end_concatenation'
 require_relative 'rubocop/cop/style/method_call_without_args_parentheses'
 require_relative 'rubocop/cop/style/method_call_with_args_parentheses'
@@ -529,6 +528,7 @@ require_relative 'rubocop/cop/style/string_hash_keys'
 require_relative 'rubocop/cop/style/string_literals'
 require_relative 'rubocop/cop/style/string_literals_in_interpolation'
 require_relative 'rubocop/cop/style/string_methods'
+require_relative 'rubocop/cop/style/strip'
 require_relative 'rubocop/cop/style/struct_inheritance'
 require_relative 'rubocop/cop/style/symbol_array'
 require_relative 'rubocop/cop/style/symbol_literal'
