+++ lib/rubocop/cop/style/strip.rb
