+++ spec/rubocop/cop/style/strip_spec.rb
